01_source_info/requirements.txt
aiosignal==1.3.1
googleapis-common-protos==1.63.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pandas==2.1.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
yfinance==0.2.40
